+++ p2panda-store-next/src/operations/sqlite.rs
@@ -36,12 +36,14 @@ where
                             payload_size,
                             payload_hash,
                             timestamp,
+                            seq_num,
                             header,
+                            header_size,
                             body,
                             extensions
                         )
                     VALUES
-                        (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+                        (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
                     ",
                 )
                 .bind(id.to_hex())
@@ -55,10 +57,12 @@ where
                 .bind(operation.header.payload_size.to_string())
                 .bind(operation.header.payload_hash.map(|hash| hash.to_hex()))
                 .bind(operation.header.timestamp.to_string())
+                .bind(operation.header.seq_num.to_string())
                 .bind(
                     encode_cbor(&operation.header)
                         .map_err(|err| SqliteError::Encode("header".to_string(), err))?,
                 )
+                .bind(operation.header.to_bytes().len().to_string())
                 .bind(operation.body.map(|body| body.to_bytes()))
                 .bind(
                     encode_cbor(&operation.header.extensions)
@@ -69,6 +73,7 @@ where
                 .map_err(SqliteError::Sqlite)
             })
             .await?;
+
         Ok(result.rows_affected() > 0)
     }
 
@@ -119,6 +124,7 @@ where
                 .map_err(SqliteError::Sqlite)
             })
             .await?;
+
         Ok(result.is_some())
     }
 
@@ -139,15 +145,34 @@ where
                 .map_err(SqliteError::Sqlite)
             })
             .await?;
+
+        Ok(result.rows_affected() > 0)
+    }
+
+    async fn delete_operation_payload(&self, id: &Hash) -> Result<bool, Self::Error> {
+        let result = query(
+            "
+            UPDATE
+                operations_v1
+            SET
+                body = NULL
+            WHERE
+                operations_v1.hash = ?
+            ",
+        )
+        .bind(id.to_hex())
+        .execute(&self.pool)
+        .await?;
+
         Ok(result.rows_affected() > 0)
     }
 }
 
 /// Single operation row as it is inserted in the SQLite database.
-#[derive(Debug, FromRow)]
-struct OperationRow {
+#[derive(Clone, Debug, FromRow)]
+pub(crate) struct OperationRow {
     hash: String,
-    header: Vec<u8>,
+    pub(crate) header: Vec<u8>,
     body: Option<Vec<u8>>,
 }